Add a test for mondodb version 5

+++ tests/packages.sh
@@ -91,7 +91,7 @@ fi
 if [ ${PIPELINE_ID} -eq "3" ]; then
 	test_header "MongoDB 3.4"
 	export MONGODB_PORT="27018"
-	export MONGODB_VERSION="3.4.19"
+	export MONGODB_VERSION="3.4.24"
 	bash packages/mongodb.sh
 	netstat -lnp | grep "${MONGODB_PORT}.*mongod"
 	kill "$(cat ${HOME}/mongodb/mongod.lock)"
@@ -99,7 +99,7 @@ if [ ${PIPELINE_ID} -eq "3" ]; then
 	rm -rf "${HOME}/mongodb/"
 
 	export MONGODB_PORT="27018"
-	export MONGODB_VERSION="3.4.19"
+	export MONGODB_VERSION="3.4.24"
 	export MONGODB_STORAGE_ENGINE="mmapv1"
 	bash packages/mongodb.sh
 	netstat -lnp | grep "${MONGODB_PORT}.*mongod"
@@ -109,16 +109,31 @@ if [ ${PIPELINE_ID} -eq "3" ]; then
 
 	test_header "MongoDB 3.6"
 	export MONGODB_PORT="27018"
-	export MONGODB_VERSION="3.6.10"
+	export MONGODB_VERSION="3.6.23"
 	bash packages/mongodb.sh
 	netstat -lnp | grep "${MONGODB_PORT}.*mongod"
 	kill "$(cat ${HOME}/mongodb/mongod.lock)"
 	sleep 5
 	rm -rf "${HOME}/mongodb/"
+fi
 
-	test_header "MongoDB 4.0"
+# MongoDB version 4
+if [ ${PIPELINE_ID} -eq "1" ]; then
+	test_header "MongoDB 4.4"
 	export MONGODB_PORT="27018"
-	export MONGODB_VERSION="4.0.6"
+	export MONGODB_VERSION="4.4.15"
+	bash packages/mongodb.sh
+	netstat -lnp | grep "${MONGODB_PORT}.*mongod"
+	kill "$(cat ${HOME}/mongodb/mongod.lock)"
+	sleep 5
+	rm -rf "${HOME}/mongodb/"
+fi
+	
+# MongoDB version 5
+if [ ${PIPELINE_ID} -eq "2" ]; then
+	test_header "MongoDB 5.0"
+	export MONGODB_PORT="27018"
+	export MONGODB_VERSION="5.0.9"
 	bash packages/mongodb.sh
 	netstat -lnp | grep "${MONGODB_PORT}.*mongod"
 	kill "$(cat ${HOME}/mongodb/mongod.lock)"
